%% and (XY)
Turns out the (XY) and (X),Y address modes are overkill.  XY is equivalent to (zp),Y and thus sufficiently powerful.  Created an X address mode for those cases where Y is the loop index but not needed for addressing.

Refactored the subroutine parameters to use a new %%n and %%n.k syntax, which maps to $Dn00-$D800, using layers to keep parameters from clashing.